Added a new `getHtmlString` callback.
Added a new `getHtmlString` callback on the application.  It returns the
stringified version of the vdom at that current state.  Could be used on,
say, a node backend to generate a template to push out.  Soon-coming
should add the ability to hot-load html into an existing DOM instead of
wiping and re-rendering the nodes, which can be done for now.